+++ sumo/src/netload/NLTriggerBuilder.cpp
@@ -118,7 +118,7 @@ NLTriggerBuilder::buildTrigger(MSNet &net,
             unsigned int count = helper.getInt(attrs, SUMO_ATTR_POS);
             float duration = helper.getFloat(attrs, SUMO_ATTR_TO);
             float deviation  = helper.getFloat(attrs, SUMO_ATTR_XTO);
-            unsigned int entering  = helper.getFloat(attrs, SUMO_ATTR_ENTERING);
+            unsigned int entering  = (unsigned int) helper.getFloat(attrs, SUMO_ATTR_ENTERING);
             //insert in MSPhoneNet
             MSPhoneNet* pPhone = MSNet::getInstance()->getMSPhoneNet();
             if (pPhone->getMSPhoneCell(cell_id) != 0)
